+++ webviz_core_components/wrapped_components/select_with_label.py
@@ -12,6 +12,12 @@ class SelectWithLabel(html.Div):
     - label (string; optional):
         The text of the label
 
+    - collapsible (bool; optional):
+        Wraps the select in a collapsible box
+
+    - open_details (bool; default True):
+        Used to set initial opened/closed state if the collapsible box is used
+
     - wrapper_id (string; optional):
         Id of the wrapping div
 
@@ -33,6 +39,8 @@ class SelectWithLabel(html.Div):
     def __init__(
         self,
         label: str = None,
+        collapsible: bool = False,
+        open_details: bool = True,
         wrapper_id: str = None,
         persistence: bool = True,
         persistence_type: str = "session",
@@ -41,7 +49,10 @@ class SelectWithLabel(html.Div):
         super().__init__()
         if wrapper_id is not None:
             self.id = wrapper_id
-        children = [html.Label(label)] if label else []
+        if collapsible:
+            children = [html.Summary(label)] if label else []
+        else:
+            children = [html.Label(label)] if label else []
         children.append(
             BaseSelect(
                 persistence=persistence,
@@ -49,4 +60,10 @@ class SelectWithLabel(html.Div):
                 **kwargs,
             )
         )
-        self.children = html.Div(style={"fontSize": "15px"}, children=children)
+        if collapsible:
+            self.children = html.Div(
+                style={"fontSize": "15px"},
+                children=html.Details(open=open_details, children=children),
+            )
+        else:
+            self.children = html.Div(style={"fontSize": "15px"}, children=children)